agents/pyproject.toml
[project]
name = "booking-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "firebase-admin>=7.4.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "langgraph-checkpoint-postgres>=3.0.5",
    "nest-asyncio>=1.6.0",
    "paho-mqtt>=2.1.0",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.13.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "supabase>=2.29.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.44.0",
]
